python/requirements.txt
# Python dependencies for audio analysis
librosa
dtw-python
numpy
scipy
soundfile 